frontend: remove portfolio and merge into dashboard page
