+++ frontend/src/App.tsx
@@ -12,7 +12,6 @@ const DashboardPage      = lazy(() => import('@/pages/DashboardPage'))
 const MarketOverviewPage = lazy(() => import('@/pages/MarketOverviewPage'))
 const ExchangeDetailPage = lazy(() => import('@/pages/ExchangeDetailPage'))
 const AssetDetailPage    = lazy(() => import('@/pages/AssetDetailPage'))
-const PortfolioPage      = lazy(() => import('@/pages/PortfolioPage'))
 const WalletDetailPage   = lazy(() => import('@/pages/WalletDetailPage'))
 
 function PageLoader() {
@@ -40,11 +39,7 @@ export default function App() {
               <Route path="/market" element={<MarketOverviewPage />} />
               <Route path="/market/:exchangeCode" element={<ExchangeDetailPage />} />
               <Route path="/market/:exchangeCode/:ticker" element={<AssetDetailPage />} />
-              <Route path="/portfolio" element={<PortfolioPage />} />
               <Route path="/wallets/:currencyCode" element={<WalletDetailPage />} />
-              {/* Legacy redirects */}
-              <Route path="/orders" element={<Navigate to="/portfolio?tab=orders" replace />} />
-              <Route path="/trades" element={<Navigate to="/portfolio?tab=trades" replace />} />
             </Route>
 
             <Route path="*" element={<Navigate to="/dashboard" replace />} />
